trips-backend/requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
